Remove Cloudflare chat relay infrastructure
