+++ .env.example
@@ -55,10 +55,6 @@ INTERNAL_API_SECRET
 INTERNAL_SERVICE_AUTH_SECRET
 API_KEYS_ENABLED
 WEBHOOKS_ENABLED
-CHAT_STREAM_RELAY_SECRET
-NEXT_PUBLIC_CHAT_STREAM_RELAY_URL
-# Dev only: set to true to use the relay while running `npm run chat-stream:dev`
-NEXT_PUBLIC_CHAT_STREAM_RELAY_LOCAL
 PROVIDER_KEYS_SECRET
 # Required (32+ chars) to store MCP servers that use bearer/header auth
 MCP_CREDENTIAL_ENCRYPTION_KEY
