+++ config/on-prem-convex-runtime-baseline.json
@@ -52,7 +52,6 @@
     "src/app/api/v1/conversations/route.ts",
     "src/app/api/v1/conversations/share/route.ts",
     "src/app/api/v1/conversations/stop/route.ts",
-    "src/app/api/v1/conversations/stream-auth/route.ts",
     "src/app/api/v1/daytona/run/route.ts",
     "src/app/api/v1/discovery/route.ts",
     "src/app/api/v1/extensions/[extensionId]/[...path]/route.ts",
