+++ docs/api-reference/workflows/conversations.mdx
@@ -10,7 +10,7 @@ Typical flow:
 1. Create or load a conversation with `/api/v1/conversations`.
 2. Add user-visible messages with `/api/v1/conversations/message`.
 3. Run an Act turn with `/api/v1/conversations/act`.
-4. Verify stream resume credentials with `/api/v1/conversations/stream-auth` when the client reconnects after a disconnect.
+4. On reconnect, reload the authoritative conversation state; persistent generations continue through the server-side stream reader.
 5. Stop an active turn with `/api/v1/conversations/stop`.
 
 Conversation messages may contain text, file references, generated UI data, and tool invocation parts. Keep generated and persisted message shapes aligned with `@overlay/app-core`.
