+++ docs/changelog.mdx
@@ -6,7 +6,7 @@ description: "Overlay Web documentation and API changes."
 ## July 2026
 
 - **AWS/Postgres UI parity** — Self-hosted deployments now load BYOK models in chat, accept same-origin browser mutations behind a configured HTTPS reverse proxy, expose Files independently of knowledge search, complete onboarding against available capabilities, and hide unavailable connectors.
-- **Self-hosted automation streaming** — Scheduled and manual Postgres automation runs now use direct stream persistence instead of requiring the managed Cloudflare stream mirror.
+- **Self-hosted automation streaming** — Scheduled and manual Postgres automation runs now use direct stream persistence with disconnect-safe server draining.
 - **AWS-native S3 credentials** — EC2 and ECS deployments can use bucket-scoped workload roles through the AWS SDK credential chain instead of storing long-lived S3 access keys.
 - **Docs pipeline restored** — `npm run docs:check` now passes. Added missing API boundary definitions for `/api/v1/admin/*` routes and the generated OpenAPI reference.
 - **Enterprise security launch checklist** — Added `docs/deploy-operate/enterprise-security-launch-checklist.mdx` with cloud-side controls for Vercel, R2, Daytona, Convex, WorkOS, Stripe, and security telemetry.
