+++ docs/openapi/overlay-web.openapi.json
@@ -2138,13 +2138,6 @@
                     "exclusiveMinimum": true,
                     "minimum": 0
                   },
-                  "streamPersistenceMode": {
-                    "type": "string",
-                    "enum": [
-                      "cloudflare-mirror",
-                      "direct"
-                    ]
-                  },
                   "mentions": {
                     "type": "array",
                     "items": {
@@ -2347,13 +2340,6 @@
                     "type": "number",
                     "exclusiveMinimum": true,
                     "minimum": 0
-                  },
-                  "streamPersistenceMode": {
-                    "type": "string",
-                    "enum": [
-                      "cloudflare-mirror",
-                      "direct"
-                    ]
                   },
                   "mentions": {
                     "type": "array",
@@ -2845,90 +2831,6 @@
                 "required": [
                   "conversationId"
                 ],
-                "additionalProperties": true
-              }
-            }
-          }
-        }
-      }
-    },
-    "/api/v1/conversations/stream-auth": {
-      "post": {
-        "operationId": "postConversationsStreamAuth",
-        "summary": "Create a stream auth token",
-        "tags": [
-          "Conversations"
-        ],
-        "security": [
-          {
-            "browserSession": []
-          },
-          {
-            "bearerAuth": []
-          }
-        ],
-        "responses": {
-          "200": {
-            "description": "Successful response. Stable response schemas are documented in workflow guides when available.",
-            "content": {
-              "application/json": {
-                "schema": {
-                  "description": "Response shape is route-specific and may be backed by @overlay/app-core contracts."
-                }
-              }
-            }
-          },
-          "400": {
-            "description": "Invalid request.",
-            "content": {
-              "application/json": {
-                "schema": {
-                  "$ref": "#/components/schemas/ErrorResponse"
-                }
-              }
-            }
-          },
-          "401": {
-            "description": "Missing or invalid authentication.",
-            "content": {
-              "application/json": {
-                "schema": {
-                  "$ref": "#/components/schemas/ErrorResponse"
-                }
-              }
-            }
-          },
-          "500": {
-            "description": "Unexpected server error.",
-            "content": {
-              "application/json": {
-                "schema": {
-                  "$ref": "#/components/schemas/ErrorResponse"
-                }
-              }
-            }
-          }
-        },
-        "requestBody": {
-          "required": true,
-          "content": {
-            "application/json": {
-              "schema": {
-                "type": "object",
-                "properties": {
-                  "accessToken": {
-                    "type": "string",
-                    "minLength": 1
-                  },
-                  "userId": {
-                    "type": "string",
-                    "minLength": 1
-                  },
-                  "conversationId": {
-                    "type": "string",
-                    "minLength": 1
-                  }
-                },
                 "additionalProperties": true
               }
             }
