+++ internal-docs/on-prem-parity/on-prem-parity-program.mdx
@@ -270,8 +270,8 @@ P2 makes persisted chat behavior independent of Convex:
 - Text deltas are batched into Postgres while the model streams. The server keeps a background stream consumer alive after a browser disconnect, so the aggregate message can continue to advance.
 - Refresh and second-tab recovery reload the authoritative persisted message snapshot. Local active streams are not overwritten by their own event notifications.
 - Stop locks generating rows, persists the interruption sentinel, deletes obsolete deltas, and prevents a late model completion from overwriting the stopped result. Other tabs stop their local reader when they observe the event.
-- Conversation share visibility, public-token reads, turn deletion, generated UI updates, and stream relay ownership checks all use the selected repository.
-- The optional chat-stream relay can still replay exact UI stream frames. Postgres recovery does not require that relay: without it, recovery resumes from the latest durable text/message snapshot.
+- Conversation share visibility, public-token reads, turn deletion, and generated UI updates all use the selected repository.
+- Chat streams directly from the web runtime. Recovery resumes from the latest durable text/message snapshot and does not depend on an external relay.
 
 Knowledge suggestions, notebook-agent behavior, and extension planning are not P2 chat dependencies. Core knowledge retrieval is owned by P4; notebook usage accounting and extension planning remain separately gated.
