+++ internal-docs/on-prem-parity/postgres-app-data-implementation.mdx
@@ -85,7 +85,7 @@ Use separate schemas/databases for these concerns. They have different ownership
 | Step 9 | `3df3b97f1` | Documented and enforced the `pgvector` readiness boundary. Postgres app-data v1 does not imply vector parity. |
 | Step 9.1 | `bbc973a59` | Re-enabled Postgres chat title generation through the usage policy abstraction and disabled integrations/connectors in Postgres pilot mode. |
 | P1 runtime | Current branch | Added migration locking and compatibility checks, Postgres idempotency/replay stores, Redis rate limiting, durable jobs/outbox/scheduler, worker runtime, and Postgres model-catalog persistence. |
-| P2 chat | Current branch | Added transactional conversation events, cursor-based browser synchronization, batched text-delta persistence, disconnect-safe background draining, stop/finalize race protection, repository-backed message/share/stream-auth routes, and provider-neutral public shares. |
+| P2 chat | Current branch | Added transactional conversation events, cursor-based browser synchronization, batched text-delta persistence, disconnect-safe background draining, stop/finalize race protection, repository-backed message/share routes, and provider-neutral public shares. |
 | P3a projects | Current branch | Added provider-neutral project CRUD, hierarchy locks and cycle checks, linked-resource authorization, recursive soft deletion, and Postgres project UI capability exposure. |
 | P3b files/storage | Current branch | Added atomic extracted-document writes, file hierarchy constraints, unique upload intents, durable S3 cleanup jobs, bulk duplicate reconciliation, and scheduled object reconciliation. |
 | P4a knowledge schema | Current branch | Added memories, knowledge chunks, versioned pgvector embeddings, ownership indexes, and account-deletion coverage. |
@@ -282,10 +282,10 @@ Important constraints:
 - `useLiveConversationSync` is guarded so Postgres mode does not mount Convex subscriptions.
 - `/api/v1/conversations/events` uses a durable sequence cursor and long polling. Transactional `pg_notify` wakes a shared listener without repeated idle queries; correctness still comes from the event table, not the notification.
 - The active HTTP stream remains the lowest-latency path. Text deltas are flushed to Postgres in bounded batches and the server drains the stream in the background if the browser disconnects.
-- A refresh or second tab reloads the aggregate persisted message. If the optional stream relay is configured, exact UI stream frames can also be replayed.
+- A refresh or second tab reloads the aggregate persisted message from authoritative application state.
 - `/api/v1/generate-title` is allowed in Postgres mode as a supported chat-persistence route.
 - Chat title generation uses `chatUsagePolicy`, not direct Convex billing queries.
-- Memory, tools, webhooks, and extension planning remain disabled in Postgres mode. Durable text stream recovery is supported; exact relay-frame replay is optional.
+- Memory, tools, webhooks, and extension planning remain disabled in Postgres mode. Durable text stream recovery is supported without an external relay.
 
 Usage policy:
