+++ package.json
@@ -57,7 +57,7 @@
     "test:p1:postgres:remote": "npm run app-db:contract:migrate && NODE_OPTIONS=--conditions=react-server tsx --env-file=.env.on-prem-contract-tests.local --test src/server/app-data/contracts/postgres-runtime-foundations.test.ts",
     "test:p1:redis": "npm run app-db:up && OVERLAY_REDIS_URL=redis://localhost:56379 NODE_OPTIONS=--require=./scripts/register-server-only.cjs tsx --test src/server/shared/providers/redis-rate-limiter.test.ts",
     "check:p1": "npm run test:p1:unit && npm run test:p1:postgres && npm run test:p1:redis && npm run check:on-prem-parity",
-    "test:p2:unit": "NODE_OPTIONS=--require=./scripts/register-server-only.cjs tsx --test src/server/conversations/chat-stream-persistence.test.ts src/features/chat/components/chat/chatStreamResume.test.ts",
+    "test:p2:unit": "NODE_OPTIONS=--require=./scripts/register-server-only.cjs tsx --test src/server/conversations/chat-stream-persistence.test.ts",
     "test:p2:postgres": "npm run test:app-data-contracts:postgres",
     "test:p2:postgres:remote": "npm run test:app-data-contracts:postgres:remote",
     "check:p2": "npm run test:p2:unit && npm run test:p2:postgres && npm run check:on-prem-parity && npm run test:on-prem:no-convex-browser",
@@ -187,9 +187,6 @@
     "convex:backfill:storage": "node --experimental-strip-types scripts/convex-storage-backfill.ts",
     "convex:backfill:filesystem": "node --experimental-strip-types scripts/backfill-canonical-filesystem.ts",
     "convex:reindex:knowledge": "node --experimental-strip-types scripts/convex-knowledge-reindex.ts",
-    "chat-stream:dev": "npm --prefix workers/chat-stream run dev",
-    "chat-stream:deploy": "npm --prefix workers/chat-stream run deploy",
-    "test:chat-stream-worker": "npm --prefix workers/chat-stream run typecheck",
     "test:jwt": "node --env-file=../overlay-desktop/.env.local --experimental-strip-types scripts/smoke-jwt.ts",
     "test:r2:connectivity": "node --env-file=.env.local --experimental-strip-types scripts/smoke-r2-connectivity.ts",
     "test:r2:presign": "node --env-file=.env.local --experimental-strip-types scripts/smoke-r2-presign.ts",
