+++ packages/overlay-api-client/src/chat/types.ts
@@ -66,18 +66,3 @@ export interface ConversationMessageRequest {
 }
 
 export type ActConversationRequest = Record<string, unknown>
-
-export interface StreamAuthRequest {
-  conversationId?: string
-  accessToken?: string
-  userId?: string
-  [key: string]: unknown
-}
-
-export interface StreamAuthResponse {
-  token?: string
-  streamToken?: string
-  expiresAt?: number
-  expiresIn?: number
-  [key: string]: unknown
-}
