+++ packages/overlay-api-client/src/index.ts
@@ -16,8 +16,6 @@ export type {
   ConversationQuery,
   CreateConversationRequest,
   CreateConversationResponse,
-  StreamAuthRequest,
-  StreamAuthResponse,
   UpdateConversationRequest,
 } from './chat/types'
 export type { FileQuery } from './files/types'
