--- src/app/api/v1/conversations/stream-auth/route.ts
@@ -1,7 +1,0 @@
-import type { NextRequest } from 'next/server'
-import { handleBffRoute, type BffDomainService } from '../../_utils/bff'
-import * as domainService from '@/server/app-api/v1/conversations/stream-auth/route'
-
-export async function POST(request: NextRequest) {
-  return handleBffRoute(request, {}, domainService.POST as BffDomainService)
-}
