+++ src/features/chat/components/ChatExperience.tsx
@@ -237,7 +237,6 @@ export default function ChatExperience({
     ensureConversationRuntime,
     replaceConversationRuntime,
     activeRuntime,
-    chatStreamRelayApi,
     chat0,
     chat1,
     chat2,
@@ -640,7 +639,6 @@ export default function ChatExperience({
     actChat,
     authUserId: authUser?.id,
     chatInstances,
-    chatStreamRelayApi,
     completeSession,
     convexAccessToken,
     enableConvexLiveSync: convexLiveSyncEnabled,
@@ -1536,7 +1534,6 @@ export default function ChatExperience({
     chat1,
     chat2,
     chat3,
-    chatStreamRelayApi,
     forceLiveSyncRender: onRuntimeMessagesChanged,
     getActiveRuntime: getActiveRuntimeForStop,
     isActiveLoading,
