+++ src/server/app-api/v1/conversations/act/route-helpers.ts
@@ -12,7 +12,6 @@ export const MIN_ACT_ABORT_TIMEOUT_MS = 30_000
 export const MAX_ACT_ABORT_TIMEOUT_MS = 780_000
 export const MAX_ACT_MODEL_ATTEMPTS = 5
 
-export type ActStreamPersistenceMode = 'cloudflare-mirror' | 'direct'
 export type ActModelAttemptFailureReason = 'budget' | 'pricing' | 'provider' | 'reservation'
 
 export type ActModelAttemptFailure = {
@@ -134,18 +133,6 @@ export function resolveEffectiveActModelId(modelId?: string): string {
   return isLegacyFreeTierDefaultModelId(requestedModelId)
     ? FREE_TIER_DEFAULT_MODEL_ID
     : requestedModelId
-}
-
-export function resolveActStreamPersistence(params: {
-  requestedMode?: ActStreamPersistenceMode
-}): {
-  mode: ActStreamPersistenceMode
-  useCloudflareStreamMirror: boolean
-} {
-  if (params.requestedMode === 'direct') {
-    return { mode: 'direct', useCloudflareStreamMirror: false }
-  }
-  return { mode: 'cloudflare-mirror', useCloudflareStreamMirror: true }
 }
 
 export function resolveActMultiModelState(params: {
