+++ src/shared/schemas/chat.ts
@@ -59,11 +59,6 @@ export const StopConversationRequest = z.object({
   partialParts: z.array(z.record(z.unknown())).optional(),
 }).passthrough()
 
-export const StreamAuthRequest = z.object({
-  ...AuthFields,
-  conversationId: z.string().min(1).optional(),
-}).passthrough()
-
 export const ShareConversationRequest = z.object({
   ...AuthFields,
   conversationId: z.string().min(1),
@@ -94,7 +89,6 @@ export const ActConversationRequest = z.object({
   requestedToolIds: z.unknown().optional(),
   memoryEnabled: z.boolean().optional(),
   actAbortTimeoutMs: z.number().finite().positive().optional(),
-  streamPersistenceMode: z.enum(['cloudflare-mirror', 'direct']).optional(),
   mentions: z.array(z.object({
     type: z.string(),
     id: z.string(),
