--- workers/chat-stream/package.json
@@ -1,17 +1,0 @@
-{
-  "name": "@overlay/chat-stream-worker",
-  "private": true,
-  "license": "AGPL-3.0-or-later",
-  "type": "module",
-  "scripts": {
-    "dev": "wrangler dev",
-    "deploy": "wrangler deploy",
-    "check:no-legacy-app-api": "node ./scripts/ensure-no-legacy-app-api.cjs",
-    "typecheck": "npm run check:no-legacy-app-api && tsc --noEmit"
-  },
-  "devDependencies": {
-    "@cloudflare/workers-types": "^4.20260511.1",
-    "typescript": "^5.9.3",
-    "wrangler": "^4.90.1"
-  }
-}
